requirements.txt
aiohttp
orjson
loguru
python3-xid
asyncpg
uvloop

aiofile
ffmpy3
ffmpeg-python